Update AT24.js
Remove i2ca argument from every function. It's better to make multiple eeprom objects, and that way you can have devices with different page sizes in use, and the module is smaller for the majority of use cases where a single EEPROM is in use.

Also, renamed arrayToString to aToS to reduce memory footprint by a few jsvars.