[UG96] : delay the socket opening to run in TLS
Too much time is taken in the DER decoding within the mbed code.
Then the handshake cannot be finalized .
Finaly, the remote terminates the connection before operating in secure mode.

Solution is to delay the socket opening.
Port 443 is used since reserved for conveying data on a secure socket.
Tested with TLS (require("tls").connect) but also with HTTPS (require("http").get)