[UG96] : delay the PS attachment so that it passes most of time at first attempt
Without this optimization,
PS attachment is tried immediatly after having measured a signal,
but PS attachment frequently fails (at least half of the time) at first attempt
(CME ERROR : 148 = Unspecified GPRS error)
then an automatic PLMN registration is entered which is longer
than having this optimization.
Also, the next attempts are done on cells having lowest signal strength and that
finally registration may occur on a PLMN which is not the best available on the field.

To summarize, this is an otimization to be PS attach quickly on the best cell most of the time.