Add files via upload
Thr 2018.11.01

Two modules are generic in nature. Although there isn't a great need for these outside of Neopixel use, they may be useful for color LCD displays.

If it is decided that modules Color and Colors should remain specific to a suite of Neopixel modules, such as NeopixelColor and NeopixelColors then I will rename and resubmit.

It would be nice to have a tutorial folder, Neopixel, to deposit examples and tutorials.